Use poetry cache. (#7)
Able to partially take advantage of the poetry github dependency cache. For reasons not understood, tests under version Python 3.9 were failing because they were being run under 3.8.

Fixed release job so that it refers to the release version via a workflow environment variable rather than a file. This should allow it to refer to the correct file when uploading the python package to a github release.